refactor: unify domain package path
