+++ domain/src/main/java/nexters/dividend/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package nexters.domain.base;
+package nexters.dividend.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
