+++ domain/src/main/java/nexters/dividend/domain/stock/Stock.java
@@ -4,11 +4,10 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import nexters.domain.base.BaseEntity;
+import nexters.dividend.domain.BaseEntity;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 @Entity
 @Getter
